feat(client): Bevy 0.16 → 0.17 (upgrade step 3)
Third step of the Bevy upgrade. 0.17's big change is the first-party crate split
(types moved out of `bevy::render::*` into `bevy::mesh`, `bevy::camera`,
`bevy::light`, `bevy::shader`, `bevy::asset`, `bevy::post_process`). Compiler-
driven migration of `atomr-worlds-client`:

- Module relocations: `RenderAssetUsages` → `bevy::asset`; `Mesh`/`Indices`/
  `PrimitiveTopology`/`MeshVertexBufferLayoutRef` → `bevy::mesh`;
  `RenderTarget`/`ClearColorConfig`/`Exposure` → `bevy::camera`;
  `NoFrustumCulling` → `bevy::camera::visibility`; `NotShadowCaster`/
  `NotShadowReceiver`/`CascadeShadowConfig*` → `bevy::light`; `ShaderRef` →
  `bevy::shader`; `Bloom` → `bevy::post_process::bloom` (adds the
  `bevy_post_process` feature).
- `Query::get_single[_mut]` removed → `single[_mut]` (returns `Result`).
- `EventWriter`/`MessageWriter::send` → `write`.
- `EntityCommands::despawn_recursive` → `despawn` (now recursive by default).
- `Camera.hdr` field → the `Hdr` component (`bevy::render::view::Hdr`).
- `Window.cursor_options` field → a separate `CursorOptions` component (query it).
- `Material::specialize`: `MaterialPipeline<Self>` → `MaterialPipeline`.
- `RenderDevice::poll(Maintain::Wait)` → `poll(PollType::Wait)` (returns Result).
- `WindowResolution::new` takes `u32` physical pixels (was `f32`).
- `InputSystem` set → `InputSystems`; bevy feature `zstd` → `zstd_rust`.
- MSRV 1.85 → 1.88.

Verified: `cargo test --workspace` → 614 passed, 0 failed.

+++ crates/atomr-worlds-client/src/main.rs
@@ -85,7 +85,8 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
             // scale_factor_override(1.0) so logical pixels == physical
             // pixels. Without this, a HiDPI display (scale 2) would
             // create a 2x larger swapchain than the capture path expects.
-            WindowResolution::new(scenario.width as f32, scenario.height as f32)
+            // Bevy 0.17: WindowResolution::new takes physical pixels as u32.
+            WindowResolution::new(scenario.width as u32, scenario.height as u32)
                 .with_scale_factor_override(1.0),
         ),
         None => (ViewMode::Fp, WindowResolution::default()),
